Update config files
- Update editor config with indent style and width
- Update eslint to indent on `switch`
- Updates to webapp manifest
- Add preconnect link hints
- Make icons compatible with new maskable icons spec
- Remov service worker registration function
- Update CSP
- Use eslint on entire project instead of just `js/`
