+++ .editorconfig
@@ -13,6 +13,8 @@ charset = utf-8
 indent_style = tab
 trim_trailing_whitespace = true
 max_line_length = 80
+indent_style = tab
+indent_size = 4
 
 [*.{md,markdown,yml,json}]
 indent_style = space
