+++ .eslintrc.json
@@ -4,19 +4,18 @@
         "es6": true
     },
     "extends": "eslint:recommended",
-    "parser": "babel-eslint",
     "parserOptions": {
         "sourceType": "module",
         "ecmaVersion": 2018
     },
-    "plugins": [
-      "async-await",
-      "frontmatter"
-    ],
+    "plugins": ["async-await", "frontmatter"],
     "rules": {
         "indent": [
             2,
-            "tab"
+            "tab",
+            {
+              "SwitchCase": 1
+            }
         ],
         "quotes": [
             2,
@@ -27,15 +26,9 @@
             "always"
         ],
         "no-console": 0,
-	"no-prototype-builtins": 0,
+        "no-async-promise-executor": 0,
+        "no-prototype-builtins": 0,
         "async-await/space-after-async": 2,
-        "async-await/space-after-await": 2,
-        "generator-star-spacing": ["error", {
-            "before": false,
-            "after": true,
-            "anonymous": "neither",
-            "method": {"before": true, "after": false}
-        }]
-    },
-    "globals": {}
+        "async-await/space-after-await": 2
+    }
 }
