+++ _data/icons.csv
@@ -6,6 +6,7 @@ public-share,img/adwaita-icons/places/folder-publicshare.svg
 web-browser,img/adwaita-icons/apps/web-browser.svg
 contact-new,img/adwaita-icons/actions/contact-new.svg
 avatar-default,img/adwaita-icons/status/avatar-default.svg
+system-software-install,img/adwaita-icons/apps/system-software-install.svg
 call-start,img/adwaita-icons/actions/call-start.svg
 go-next,img/adwaita-icons/actions/go-next.svg
 go-prev,img/adwaita-icons/actions/go-prev.svg
