+++ _includes/footer.html
@@ -38,6 +38,14 @@
 			</a>
 		{% endif %}
 	</div>
+	{% if site.service_worker %}
+		<button is="pwa-install" class="btn btn-primary btn-big btn-wide" src="{{ site.service_worker }}" scope="/" title="Install Web App" reloadonupdate="">
+			<svg class="icon current-color" role="img" aria-hidden="true">
+				<use xlink:href="/img/icons.svg#system-software-install"></use>
+			</svg>
+			<span>Install</span>
+		</button>
+	{% endif %}
 	<hr />
 	<div class="copyright">
 		<span>&copy;</span>
