+++ contact.html
@@ -4,6 +4,7 @@ title: Contact
 index: true
 permalink: contact/
 referrer: no-referrer
+pinned: true
 description: Contact form for Paradise Cove Lodge
 thumbnail:
   url: /img/raster/background/snow-400.jpg
