+++ cssnano.config.js
@@ -1,3 +1,4 @@
+/* eslint-env node */
 const defaultPreset = require('cssnano-preset-default');
 
 module.exports = defaultPreset({
