+++ img/icon-16.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:601af7dee1f7bc82afd7bb8a835e1d0fae6c83bb5cb8f82180a0aff8e970718d
-size 907
+oid sha256:c1645d6bda1ea374a530392b77c99360feeffc04999ca8799bf07db7f3f7e5d7
+size 553
