+++ img/icon-192.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:574e5bb0aae7da8b91fe1157d9ae15e2727dba1cc6d0a9a10c94aa7e21fc9f0e
-size 8347
+oid sha256:b7524111536ad239be0e248159154779d1ebc6797b49687f2453bea723210561
+size 6632
