+++ img/icon-32.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:ef16d18cb06c4b7f1be006b5635abb56a1291e948cc58472f661f8fbc4018dc5
-size 1352
+oid sha256:92dddda2e97eadcf3dcf3bdc792a0fca31134d49a0a60019a988021253cf8e62
+size 1170
