+++ img/icon-512.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6b9d17590e010b54084a5cc3652df30638b7185f201e41c0b3dfdcb76f900d29
-size 22583
+oid sha256:0242964685162816093b2f35b5e931878df654121856958b7534245b054eb397
+size 18746
