+++ index.html
@@ -4,6 +4,7 @@ title: Home
 background: sunrise
 index: true
 referrer: no-referrer
+pinned: true
 thumbnail:
   url: /img/raster/background/sunrise-400.jpg
 description: "Steakhouse &amp; Motel in Lake Isabella California. Paradise Cove serves choice angus beef and quality seafood. Renovated motel rooms all have view of the lake."
@@ -53,8 +54,4 @@ tags:
 		<h3 class="card-label center">Location</h3>
 		<p class="card-desc">On Highway 178 with great views of Lake Isabella, 3 miles past the Dam on the south side of the lake.</p>
 	</a>
-</section>
-<section>
-	<br />
-	{% include map.html geo=site.data.business.geo hidden=true %}
 </section>
