+++ lake-cam.html
@@ -4,6 +4,7 @@ title: Lake Isabella Cam & Conditions
 index: true
 permalink: /lakecam/
 background: lake
+pinned: true
 description: Live feed of our lake cam at Paradise Cove
 tags:
   - lake cam
