+++ manifest.json.config
@@ -8,13 +8,14 @@ permalink: manifest.json
 	"short_name": "{{ site.data.app.shortName | default: site.title}}",
 	"description": "{{ site.data.app.description | default: site.description }}",
 	"start_url": "{{ page.url | replace: 'manifest.json', '' }}",
-	"lang": "{{ site.lang }}",
+	"lang": "{{ site.data.app.lang | default: site.lang }}",
+	"dir": "{{ site.app.data.dir | default: 'ltr' }}",
+	"orientation": "{{ site.app.data.orientation | default: 'any' }}",
 	"display": "{{ site.data.app.display | default: 'standalone' }}",
 	"theme_color": "{{ site.data.app.theme.color | default: '#185a93' }}",
 	"background_color": "{{ site.data.app.theme.background | default: '#FEFEFE' }}",
-	"icons": {{ site.data.app.icons | jsonify }},
+	"icons": {{ site.data.app.icons | where_exp: "icon", "icon.purpose" | jsonify }},
 	"screenshots": {{ site.data.app.screenshots | jsonify }},
 	"features": {{ site.data.app.features | jsonify }},
-	"related_applications":  {{ site.data.app.related_applications | jsonify }},
  	"categories": {{ site.data.app.categories | jsonify }}
 }
