+++ menu.html
@@ -2,6 +2,7 @@
 layout: default
 title: Menu
 permalink: menu/
+pinned: true
 index: true
 background: restaurant
 description: Menu for Paradise Cove Lodge
