+++ netlify.toml
@@ -21,15 +21,15 @@
     Content-Security-Policy = '''
       default-src 'self' https://cdn.kernvalley.us https://unpkg.com/;
       base-uri 'self';
-      img-src 'self' https://*.githubusercontent.com/u/  https://maps.wikimedia.org/osm-intl/ https://www.tazdij.com/static/ftp/paradisecove/ftp/cameraview.jpg https://cdn.kernvalley.us/img/;
-      script-src 'self' moz-extension: https://cdn.kernvalley.us https://unpkg.com/  https://cdn.polyfill.io/v3/polyfill.min.js https://www.google-analytics.com/ https://www.googletagmanager.com https://polyfill.io/v3/ https://www.youtube.com/ https://www.youtube-nocookie.com/;
-      style-src 'self' 'unsafe-inline' https://cdn.kernvalley.us https://unpkg.com/ https://fonts.googleapis.com/ https://www.youtube.com/ https://www.youtube-nocookie.com/;
-      connect-src 'self' https://api.github.com/users/ https://cdn.kernvalley.us https://api.kernvalley.us https://unpkg.com/ https://cdn.polyfill.io/v3/ https://www.google-analytics.com/ https://www.googletagmanager.com/ https://i.imgur.com https://secure.gravatar.com https://maps.wikimedia.org/osm-intl/ https://fonts.gstatic.com/  https://fonts.googleapis.com/ https://maps.wikimedia.org/osm-intl/ https://www.tazdij.com/static/ftp/paradisecove/ftp/cameraview.jpg;
+      img-src 'self' https://*.githubusercontent.com/u/ https://maps.wikimedia.org/osm-intl/ https://www.tazdij.com/static/ftp/paradisecove/ftp/cameraview.jpg https://cdn.kernvalley.us/img/ https://i.imgur.com;
+      script-src 'self' moz-extension: https://cdn.kernvalley.us https://unpkg.com/ https://cdn.polyfill.io/ https://www.google-analytics.com/ https://www.googletagmanager.com https://www.youtube.com/ https://www.youtube-nocookie.com/;
+      style-src 'self' https://cdn.kernvalley.us https://unpkg.com/ https://fonts.googleapis.com/ https://www.youtube.com/ https://www.youtube-nocookie.com/ https://*.githubusercontent.com/u/;
+      connect-src 'self' https://api.github.com/users/ https://cdn.kernvalley.us https://api.kernvalley.us https://unpkg.com/ https://cdn.polyfill.io/ https://www.google-analytics.com/ https://www.googletagmanager.com/ https://i.imgur.com https://secure.gravatar.com https://maps.wikimedia.org/osm-intl/ https://fonts.gstatic.com/  https://fonts.googleapis.com/ https://maps.wikimedia.org/osm-intl/ https://www.tazdij.com/static/ftp/paradisecove/ftp/cameraview.jpg;
       font-src 'self' https://cdn.kernvalley.us https://fonts.gstatic.com/;
       media-src 'none';
       frame-src https://www.youtube.com https://www.youtube-nocookie.com https://disqus.com;
       object-src: 'none';
-      prefetch-src: 'self' https://unpkg.com/ https://cdn.kernvalley.us
+      prefetch-src: 'self' https://unpkg.com/ https://cdn.kernvalley.us https://api.github.com/users/
       worker-src: 'self';
       manifest-src 'self';
       form-action 'self';
