+++ package.json
@@ -44,10 +44,10 @@
     "git:fetch": "git fetch --prune --tags",
     "git:submodules": "if [ -f '.gitmodules' ]; then git submodule update --init --recursive; fi",
     "lint": "npm run lint:js && npm run lint:css",
-    "lint:js": "eslint ${npm_package_config_dir_js}",
+    "lint:js": "eslint .",
     "lint:css": "stylelint ${npm_package_config_dir_css}",
     "fix:css": "stylelint ${npm_package_config_dir_css} --fix",
-    "fix:js": "eslint ${npm_package_config_dir_js} --fix",
+    "fix:js": "eslint. --fix",
     "build": "npm run build:icons && npm run build:css && npm run build:js",
     "build:site": "JEKYLL_ENV=production bundle exec jekyll build",
     "build:site:dev": "bundle exec jekyll build --drafts --unpublished --future --profile --trace --strict_front_matter",
