+++ postcss.config.js
@@ -1,4 +1,4 @@
-/* global module, require */
+/* eslint-env node */
 module.exports = {
 	map: {inline: false},
 	plugins: [
