+++ rollup.config.js
@@ -1,3 +1,4 @@
+/* eslint-env serviceworker */
 import urlResolve from 'rollup-plugin-url-resolve';
 import {terser} from 'rollup-plugin-terser';
