+++ rooms.html
@@ -4,6 +4,7 @@ title: Rooms & Rates
 index: true
 permalink: /rooms/
 background: motel
+pinned: true
 referrer: no-referrer
 description: Rooms available starting at $85 / night
 thumbnail:
